hrms_project/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
python-dotenv>=1.0
openpyxl>=3.0
pandas>=1.5
Pillow>=9.0
python-dateutil>=2.9
